Export extract package used for extracting event log info
This package is required for parsing out events from TCG and CCEL logs.
