+++ extract/rtmr.go
@@ -12,7 +12,7 @@
 // License for the specific language governing permissions and limitations under
 // the License.
 
-package eventparse
+package extract
 
 import (
 	"bytes"
